fixed a bug where the bot wouldn't display never as  when someone hadn't grabbed any cookies.
added try except catch for charts in spotify just in case there's any
errors that i don't want to display on the console.

i have modified the way that data is represented on the charts. i have
added a legend, made the chart slightly more focussed and have changed
the colours. still need to find on making a working dark mode.